backend/flask_services/matcher_service/requirements.txt
# Core dependencies
Flask==3.0.0
Flask-CORS==4.0.0

# Lightweight ML stack
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.3

# Database
psycopg2-binary==2.9.10

# Production
python-dotenv==1.0.0
blinker==1.6.2  # Required by Flask

# Text processing
joblib==1.3.2
sentencepiece==0.1.99

# Web server
gunicorn==21.2.0
